+++ framework/CHANGELOG.md
@@ -16,6 +16,7 @@ Yii Framework 2 Change Log
 - Bug #20891: Fix `@return` annotation for `RequestParserInterface::parse()`, `JsonParser::parse()`, `Request::getBodyParams()`, and `Request::post()` (mspirkov)
 - Bug #20891: Fix `@param` annotation for `$values` in `Request::setBodyParams()` (mspirkov)
 - Bug #20891: Fix `@property` annotation for `Request::$bodyParams` (mspirkov)
+- Enh #20899: Add `isValid()` to `yii\data\Pagination` and `yii\data\Sort` to detect an out-of-range page or an undeclared sort attribute (WarLikeLaux)
 
 
 2.0.55 May 09, 2026